moved api call to update db to seperate file
